requirements.txt
python-dotenv
python-telegram-bot
google-generativeai
flask
gunicorn